# perf(ext/event): always set timeStamp to 0