@littledivy 2023-08-18
We can remove `[_skipInternalInit]` option from Event constructor, if we always set timestamp to 0.