@bartlomieju 2023-08-20
Opened https://github.com/denoland/manual/pull/718